clean (FileUploader.Commandable.cs): move the exclusivity-checks into FileUploader.Commandable.cs
