+++ Laerdal.McuMgr/Shared/FileUploading/FileUploader.Commandable.cs
@@ -1,6 +1,3 @@
-// ReSharper disable UnusedType.Global
-// ReSharper disable RedundantExtendsListEntry
-
 namespace Laerdal.McuMgr.FileUploading
 {
     public partial class FileUploader
